feat: implement GetCards handler to retrieve user card list from DynamoDB
