+++ server/internal/infra/server/server.go
@@ -1,9 +1,7 @@
 package server
 
 import (
-	"context"
 	"fmt"
-	"log/slog"
 	"net/http"
 
 	"github.com/labstack/echo/v4"
@@ -11,7 +9,6 @@ import (
 	oapimiddleware "github.com/oapi-codegen/echo-middleware"
 	"github.com/yamato0211/brachio-backend/internal/config"
 	"github.com/yamato0211/brachio-backend/internal/handler"
-	"github.com/yamato0211/brachio-backend/internal/infra/dynamo"
 )
 
 type Server struct {
@@ -36,26 +33,7 @@ func New() (*Server, error) {
 	e.Use(echomiddleware.Logger())
 	e.Use(oapimiddleware.OapiRequestValidator(swagger))
 
-	dc, err := dynamo.New(context.Background(), cfg.Dynamo)
-	if err != nil {
-		slog.Warn("failed to create dynamo client", slog.Attr{Key: "error", Value: slog.AnyValue(err)})
-	}
-
 	e.GET("/", HealthCheck)
-
-	e.GET("/users", func(c echo.Context) error {
-		type User struct {
-			ID   string `dynamo:"UserId,hash"`
-			Name string `dynamo:"Name"`
-		}
-		users := []User{}
-		err := dc.Table("Users").Scan().All(c.Request().Context(), &users)
-		if err != nil {
-			slog.Error("error!", slog.Attr{Key: "error", Value: slog.AnyValue(err)})
-			return c.JSON(http.StatusInternalServerError, err)
-		}
-		return c.JSON(http.StatusOK, users)
-	})
 
 	handler := handler.New()
 	RegisterHandlers(e, handler)
